Partially working with error on detailed scan
