+++ gomap_funcs.go
@@ -1,0 +1,65 @@
+package gomap
+
+import (
+	"encoding/binary"
+	"fmt"
+	"log"
+	"net"
+	"strings"
+)
+
+// createHostRange converts a input ip addr string to a slice of ips on the cidr
+func createHostRange(netw string) []string {
+	_, ipv4Net, err := net.ParseCIDR(netw)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	mask := binary.BigEndian.Uint32(ipv4Net.Mask)
+	start := binary.BigEndian.Uint32(ipv4Net.IP)
+	finish := (start & mask) | (mask ^ 0xffffffff)
+
+	var hosts []string
+	for i := start + 1; i <= finish-1; i++ {
+		ip := make(net.IP, 4)
+		binary.BigEndian.PutUint32(ip, i)
+		hosts = append(hosts, ip.String())
+	}
+
+	return hosts
+}
+
+// getLocalRange returns local ip range or defaults on error to most common
+func getLocalRange() string {
+	addrs, err := net.InterfaceAddrs()
+	if err != nil {
+		return "192.168.1.0/24"
+	}
+	for _, address := range addrs {
+		// check the address type and if it is not a loopback the display it
+		if ipnet, ok := address.(*net.IPNet); ok && !ipnet.IP.IsLoopback() {
+			if ipnet.IP.To4() != nil {
+				split := strings.Split(ipnet.IP.String(), ".")
+				return split[0] + "." + split[1] + "." + split[2] + ".0/24"
+			}
+		}
+	}
+	return "192.168.1.0/24"
+}
+
+// getLocalRange returns local ip range or defaults on error to most common
+func getLocalIP() (string, error) {
+	addrs, err := net.InterfaceAddrs()
+	if err != nil {
+		return "", err
+	}
+	for _, address := range addrs {
+		// check the address type and if it is not a loopback the display it
+		if ipnet, ok := address.(*net.IPNet); ok && !ipnet.IP.IsLoopback() {
+			if ipnet.IP.To4() != nil {
+				return ipnet.IP.String(), err
+			}
+		}
+	}
+	return "", fmt.Errorf("No IP Found!")
+}
